app/requirements.txt
django-phone-field==1.8.1